+++ routes/web.php
@@ -35,28 +35,42 @@ use App\Http\Controllers\ListingController;
 Route::get('/',[ListingController::class, 'index']);
 
 //Show Create Listing form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
+
+//the ->Middleware('auth') makes the route only accessible to logged in user
 
 //Store listing data
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 //Show edit forms
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Update listing data
-Route::put('listings/{listing}', [ListingController::class, 'update']);
+Route::put('listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete listing data
-Route::delete('listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
+
+//Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 //Single Listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 //Show Register/Create Form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+//->middleware('guest') this ensures that the logged in does see the register page
 
 //Create new user
 Route::post('/users', [UserController::class, 'store']);
 
 //Log user out
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+//Show login form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+//->name('login') is used so that middleware auth can find the login route (learnt for laragig traversy tutorial)
+
+// Login user
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
+
